More convergent packaging
Don't redownload rabbitmq
Don't try to create already existing virtualenv
Store a version for jgit and only re-compile if it has changed

+++ manager/package/scripts/jgit.py
@@ -1,3 +1,5 @@
+import os
+
 from manager.shared import dependencies_directory
 from manager.shared.script import ShellScript
 
@@ -5,11 +7,24 @@ from manager.shared.script import ShellScript
 class JgitPackageScript(ShellScript):
 	@classmethod
 	def get_script(cls):
+		jgit_dir = os.path.join(dependencies_directory, 'jgit')
+		jgit_version_file = os.path.join(jgit_dir, '.version')
 		return '''
 			cd /tmp
 			git clone git://github.com/LessThanThreeLabs/jgit.git
 			cd jgit
-			mvn install
-			mv org.eclipse.jgit.pgm/target/jgit %s
+			headsha=$(git rev-parse HEAD)
+			if [ -f %s ] && [ "$(cat %s)" == "$headsha" ]; then
+				echo "Newest jgit already packaged"
+			else
+				mvn install
+				mkdir -p %s
+				mv org.eclipse.jgit.pgm/target/jgit %s
+				echo $headsha > %s
+			fi
 			rm -rf /tmp/jgit
-		''' % dependencies_directory
+		''' % (jgit_version_file,
+				jgit_version_file,
+				jgit_dir,
+				jgit_dir,
+				jgit_version_file)
