+++ manager/package/scripts/rabbitmq.py
@@ -7,6 +7,8 @@ class RabbitmqPackageScript(ShellScript):
 	def get_script(cls):
 		return '''
 			cd %s
-			wget http://www.rabbitmq.com/releases/rabbitmq-server/v2.8.7/rabbitmq-server_2.8.7-1_all.deb
-			mv rabbitmq-server_2.8.7-1_all.deb rabbitmq-server.deb
+			if [ ! -f rabbitmq-server.deb ]; then
+				wget http://www.rabbitmq.com/releases/rabbitmq-server/v2.8.7/rabbitmq-server_2.8.7-1_all.deb
+				mv rabbitmq-server_2.8.7-1_all.deb rabbitmq-server.deb
+			fi
 		''' % dependencies_directory
